migrate controlpanel.ui.configuration.model.xml to core
