+++ application/tests/org.openjdk.jmc.flightrecorder.controlpanel.ui.test/src/test/java/org/openjdk/jmc/flightrecorder/controlpanel/ui/model/test/JfrControlTestCase.java
@@ -35,7 +35,7 @@ package org.openjdk.jmc.flightrecorder.controlpanel.ui.model.test;
 import java.io.InputStream;
 
 import org.openjdk.jmc.flightrecorder.configuration.events.IEventConfiguration;
-import org.openjdk.jmc.flightrecorder.controlpanel.ui.configuration.model.xml.XMLModel;
+import org.openjdk.jmc.flightrecorder.configuration.model.xml.XMLModel;
 import org.openjdk.jmc.flightrecorder.controlpanel.ui.model.EventConfiguration;
 import org.openjdk.jmc.rjmx.services.jfr.test.JfrTestCase;
